requirements.txt
aiogram>=3.10,<4
pydantic-settings>=2.0
google-generativeai>=0.8
moviepy>=2.0
Pillow>=10.0
numpy
aiosqlite>=0.19
rembg>=2.0
aiohttp>=3.9
